requirements.txt
# Streamlit UI
streamlit>=1.40.0

# データ分析
pandas>=2.2.0
numpy>=1.26.0

# チャート
plotly>=5.24.0

# HTTP / API
requests>=2.32.0
httpx>=0.27.0

# スクレイピング
beautifulsoup4>=4.12.0
lxml>=5.3.0

# PDF処理
PyMuPDF>=1.24.0

# AI (Gemini)
google-genai>=1.0.0

# DB
sqlalchemy>=2.0.0

# テクニカル分析
ta>=0.11.0

# 環境変数
python-dotenv>=1.0.0

# テスト
pytest>=8.0.0
pytest-asyncio>=0.24.0
